docs(examples): update proxy example to async/await
